Gestion de modification de logo de l'etablissement

+++ Backend/edt/views/viewsEtablissement.py
@@ -61,25 +61,36 @@ class EtablissementView(APIView):
         donnees = request.data
         ancienLogo=etablissement.logo
         nouveauLogo = request.FILES.get('logo')
-
+        if not nouveauLogo:
+            nouveauLogo=request.data.get('logo')
         if nouveauLogo:
-            dossier = os.path.join(settings.MEDIA_ROOT, 'etablissements')
-            os.makedirs(dossier, exist_ok=True)
-            chemin_fichier = os.path.join(dossier, nouveauLogo.name)
+            v=True
+            if ancienLogo:
+                absAncienLogo=request.build_absolute_uri(settings.MEDIA_URL + ancienLogo)
+                print(absAncienLogo)
 
-            with open(chemin_fichier, 'wb+') as destination:
-                for c in nouveauLogo.chunks():
-                    destination.write(c)
+                v=(absAncienLogo!=nouveauLogo)
+            print(v)
+            print(nouveauLogo)
+            if v:
+                dossier = os.path.join(settings.MEDIA_ROOT, 'etablissements')
+                os.makedirs(dossier, exist_ok=True)
+                chemin_fichier = os.path.join(dossier, nouveauLogo.name)
 
-            logoChemin = f"etablissements/{nouveauLogo.name}"
-
-            if ancienLogo and logoChemin!=ancienLogo and ancienLogo:
-                cheminAncienLogo=os.path.join(settings.MEDIA_ROOT, ancienLogo)
-                existeAutre = Etablissement.objects.filter(logo=ancienLogo).exclude(pk=etablissement.numEtablissement).exists()
-                if os.path.exists(cheminAncienLogo) and not existeAutre:
-                    os.remove(cheminAncienLogo)
-            donnees['logo'] = logoChemin
-
+                with open(chemin_fichier, 'wb+') as destination:
+                    for c in nouveauLogo.chunks():
+                        destination.write(c)
+                logoChemin = f"etablissements/{nouveauLogo.name}"
+                if ancienLogo and logoChemin!=ancienLogo and ancienLogo:
+                    cheminAncienLogo=os.path.join(settings.MEDIA_ROOT, ancienLogo)
+                    existeAutre = Etablissement.objects.filter(logo=ancienLogo).exclude(pk=etablissement.numEtablissement).exists()
+                    if os.path.exists(cheminAncienLogo) and not existeAutre:
+                        os.remove(cheminAncienLogo)
+                donnees['logo'] = logoChemin
+            else:
+                donnees['logo'] = ancienLogo
+        print(donnees)
+        print(donnees.get('logo'))
         serializer = EtablissementSerializer(etablissement, data=donnees)
         if serializer.is_valid():
             etablissements = serializer.save()
